mid: 324 msg: Merge de la branca 2.4
git-svn-id: svn://svn.code.sf.net/p/gehelium/code/helium/branches/helium-2.5@541 e3271b2b-fd41-4d56-bedf-4b5636213925

+++ helium-core/src/main/java/net/conselldemallorca/helium/core/model/service/DocumentHelper.java
@@ -72,8 +72,11 @@ public class DocumentHelper {
 			String arxiuNom,
 			byte[] arxiuContingut,
 			boolean isAdjunt) {
+		DocumentStore documentStore = null;
 		Long documentStoreId = getDocumentStoreIdDeVariableJbpm(taskInstanceId, processInstanceId, documentCodi);
-		if (documentStoreId == null) {
+		if (documentStoreId != null)
+			documentStore = documentStoreDao.getById(documentStoreId, false);
+		if (documentStore == null) {
 			// Si el document no existeix el crea
 			documentStoreId = documentStoreDao.create(
 					processInstanceId,
@@ -92,7 +95,6 @@ public class DocumentHelper {
 					arxiuNom,
 					(pluginGestioDocumentalDao.isGestioDocumentalActiu()) ? null : arxiuContingut,
 					documentNom);
-			DocumentStore documentStore = documentStoreDao.getById(documentStoreId, false);
 			if (arxiuContingut != null && pluginGestioDocumentalDao.isGestioDocumentalActiu())
 				pluginGestioDocumentalDao.deleteDocument(documentStore.getReferenciaFont());
 		}
@@ -156,7 +158,8 @@ public class DocumentHelper {
 				jbpmDao.deleteTaskInstanceVariable(
 						taskInstanceId,
 						PREFIX_SIGNATURA + documentCodi);
-			} else {
+			}
+			if (processInstanceId != null) {
 				jbpmDao.deleteProcessInstanceVariable(
 						processInstanceId,
 						getVarPerDocumentCodi(documentCodi, false));
@@ -236,7 +239,8 @@ public class DocumentHelper {
 			String taskInstanceId,
 			String processInstanceId,
 			String documentCodi,
-			boolean perSignarEnTasca) {
+			boolean perSignarEnTasca,
+			boolean ambInfoPsigna) {
 		Long documentStoreId = getDocumentStoreIdDeVariableJbpm(taskInstanceId, processInstanceId, documentCodi);
 		if (documentStoreId != null) {
 			DocumentDto dto = toDocumentDto(
@@ -404,19 +408,18 @@ public class DocumentHelper {
 			String taskInstanceId,
 			String processInstanceId,
 			String documentCodi) {
+		Object value = null;
 		if (taskInstanceId != null) {
-			return (Long)jbpmDao.getTaskInstanceVariable(
-			taskInstanceId,
-			getVarPerDocumentCodi(documentCodi, false));
-			/*JbpmTask taskInstance = jbpmDao.getTaskById(taskInstanceId);
-			return (Long)jbpmDao.getProcessInstanceVariable(
-					taskInstance.getProcessInstanceId(),
-					getVarPerDocumentCodi(documentCodi));*/
-		} else {
-			return (Long)jbpmDao.getProcessInstanceVariable(
+			value = jbpmDao.getTaskInstanceVariable(
+					taskInstanceId,
+					getVarPerDocumentCodi(documentCodi, false));
+		}
+		if (value == null && processInstanceId != null) {
+			value = jbpmDao.getProcessInstanceVariable(
 					processInstanceId,
 					getVarPerDocumentCodi(documentCodi, false));
 		}
+		return (Long)value;
 	}
 	
 	private Document getDocumentDisseny(
